requirements.txt
streamlit
sqlite3
whisper
google-generativeai
chromadb
sentence-transformers
